backend/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-JWT-Extended>=4.6,<5
Flask-Migrate>=4.0,<5
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
gunicorn>=22.0,<23
pytest>=8.0,<9
